plug-ins: fix uninitialized warnings.
This fixes various similar warnings:

> file-xmc.c:2276:53: warning: ‘save_rgn.height’ may be used uninitialized in this function

Note that I have not looked in details, but it would seem that these are
actually initialized by get_cropped_region() which uses goto-s quite
extensively, so obviously it's hard to follow (for humans as well as for
the compiler). So let's just assume there is no code error and let's
just add default initialization to silence the compiler.