libgimp: replace GtkFlowbox use by GtkGrid in generated metadata frame.
It is a bit more flexible. Also this fixes the ugly focus issue we had
on the comment text input (which might disappear soon anyway, but since
this frame is meant to also display user-created widgets, better to not
have a container breaking text widgets).

I am not 100% happy for the generated layout, but this is meant to
evolve anyway.