menus: add a version-specific issues link to the GIMP gitLab issues
The idea is that this may make it easier for users to spot whether an issue the have encountered is already
known, like the infamous export crash in 2.99.4. That's also why the link being used shows all reports: open
and closed ones.

The current state would work fine for releases, provided that issues have a correspnding version label. For
this reason, we should make sure that version labels are not replaced, but added to, unless a specific issues
is indeed not present in an intermediate release inbetween some others.

The naive approach of simply using (gimp-version) for the label fails for development versions in between
releases, as e.g. we have no issues with a '2.99.5' label.