devel-docs: quick command to get the right flatpak runtime version.
For debugging the flatpak, we often ask people to install the SDK and
debug data. Yet there might be several branches of the GNOME SDK
installed at once, hence flatpak will ask which version to install.

This quick command can be copy-pasted as it's a way to detect which
runtime is being used by your flatpak-ed GIMP.