plug-ins: set the "Save GeoTIFF data" checkbox insensitive when…
… there are no such data available.
No need to have people test and wonder what it changes when it makes no
difference. Also adding a tooltip to make the "present on import"
condition a bit more obvious (hopefully).

This is only on `gimp-2-10` for now, for the 2.10.24 release. I will
port this to `master` branch later.