MAINTAINERS: update.
According to GNOME docs, it looks like file is uneeded nowadays, in
favor of the doap file instead. Anyway as it's a bit easier to read and
to find for random people browsing the source tree, so let's keep it and
add myself.

Cf. https://wiki.gnome.org/MaintainersCorner#maintainers