build: remove *.rs file association with SUN Raster images
Nowadays .rs is the extension for the Rust programming language files,
and it's confusing that GIMP is trying to associate with them.

A simple rename of existing .rs images to .ras will allow them to be
opened again.

Note by reviewer: ideally file association should not rely on filename
extension, and should be detected properly (i.e. file "magic"). This way
even extension clash would not be a problem (format would be detected
whatever the extension used). Unfortunately it's apparently not the case
on Windows.
Anyway since nowadays chances to see a Rust code file are likely much
higher than seeing a Sun Raster image file, let's just accept this patch
and drop association of `.rs` on Windows.